moving things around in voxel functions
